requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
blas=1.0=mkl
brotli=1.0.9=h166bdaf_8
brotli-bin=1.0.9=h166bdaf_8
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=hd590300_0
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py311ha3edf6b_0
cryptography=39.0.1=py311h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
deprecated=1.2.13=pyh6c4a22f_0
eigen=3.4.0=h4bd325d_0
einops=0.6.1=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
ffmpeg=4.3=hf484d3e_0
fftw=3.3.10=nompi_hc118613_107
filelock=3.9.0=py311h06a4308_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.39.4=py311h459d7ec_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
hdf5=1.12.2=nompi_h2386368_101
humanize=4.6.0=pyhd8ed1ab_0
icu=58.2=hf484d3e_1000
idna=3.4=py311h06a4308_0
importlib-metadata=6.6.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46305
jinja2=3.1.2=py311h06a4308_0
jpeg=9e=h5eee18b_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py311h4dd048b_1
krb5=1.19.3=h3790be6_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=4.0.0=h27087fc_0
libaec=1.0.6=hcb278e6_1
libbrotlicommon=1.0.9=h166bdaf_8
libbrotlidec=1.0.9=h166bdaf_8
libbrotlienc=1.0.9=h166bdaf_8
libclang=10.0.1=default_hb85057a_2
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcurl=7.87.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.14=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libitk=5.3.0=hcedbc38_0
libllvm10=10.0.1=he513fc3_3
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libsqlite=3.42.0=h2797004_0
libssh2=1.10.0=haa6b8db_3
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.19.0=h5eee18b_0
libtiff=4.4.0=h82bc61c_5
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h0b41bf4_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=16.0.4=h4dfa4b3_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.7.1=py311h38be061_0
matplotlib-base=3.7.1=py311h8597a09_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py311h06a4308_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py311h06a4308_1
nibabel=5.1.0=py311h38be061_2
nspr=4.35=h27087fc_0
nss=3.89=he45b914_0
numpy=1.24.3=py311h08b1b3b_1
numpy-base=1.24.3=py311hf175353_1
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h0b41bf4_0
packaging=23.1=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pillow=9.4.0=py311h6a678d5_0
pip=23.0.1=py311h06a4308_0
ply=3.11=py_1
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py311h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.3=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.11=2_cp311
pytorch=2.0.1=py3.11_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.29.0=py311h06a4308_0
scipy=1.9.3=py311h08b1b3b_2
setuptools=66.0.0=py311h06a4308_0
simpleitk=2.2.1=py311hba2efa8_1
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=2.0.2=py311_cu118
torchio=0.18.91=pyhd8ed1ab_0
torchtriton=2.0.0=py311
torchvision=0.15.2=py311_cu118
tornado=6.3.2=py311h459d7ec_0
tqdm=4.65.0=pyhd8ed1ab_1
typing_extensions=4.5.0=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py311h06a4308_0
wheel=0.38.4=py311h06a4308_0
wrapt=1.15.0=py311h2582759_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.4.2=h5eee18b_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.5=hc292b87_0
